Remove Bukkit references. [See full message]
In light of recent uncertainty about Minecraft modding, I am removing all
references to the Bukkit project from this repository, so that it may act
as a backend for any implementation. Implementations will be added to
separate repositories, so that in the event of a takedown only the
particular implementation is lost.